Refactor error modules into smaller units
